exclude __tests__ and __mocks__ from coverage
